fix: moved FilterState and flattenKeys to avoid circular dependency
